Auth shit

+++ src/components/CommentsSection/CommentsSection.jsx
@@ -1,15 +1,20 @@
 import { Button, Textarea } from "@nextui-org/react";
 import { useSession, signIn } from "next-auth/react";
-import { useRouter } from "next/router";
-import { useState } from "react";
 import axios from "axios";
-
+import useSWR, { mutate } from "swr";
+import { useState } from "react";
 import Comment from "../Comments/Comment";
+import { AiFillDelete, AiFillEdit } from "react-icons/ai";
+import Popup from "../Popover/Popover";
+import { CiMenuKebab } from "react-icons/ci";
 
 export default function CommentsSection({ post, onNewComment }) {
-  const router = useRouter();
   const { data: session } = useSession();
-  const showForm = session;
+  const [deletingCommentId, setDeletingCommentId] = useState(null);
+  const [editMode, setEditMode] = useState(false);
+
+  const fetcher = (...args) => fetch(...args).then((res) => res.json());
+  const { data: comments } = useSWR(`/api/comment?postId=${post.id}`, fetcher);
 
   const handleSubmit = (e) => {
     e.preventDefault();
@@ -32,12 +37,57 @@ export default function CommentsSection({ post, onNewComment }) {
   };
 
   async function handleDelete(id) {
+    if (!session) {
+      alert("You need to be signed in to delete a comment.");
+      return;
+    }
+
+    setDeletingCommentId(id);
+
     try {
       await axios.delete(`/api/comment/${id}`);
       onNewComment();
+
+      mutate(`/api/comment?postId=${post.id}`, (cachedData) => {
+        return cachedData.filter((comment) => comment.id !== id);
+      });
     } catch (error) {
+      setDeletingCommentId(null);
       console.log(error);
+      return <div>Error loading comments</div>;
     }
+
+    if (!comments) {
+      return <div>Loading comments...</div>;
+    }
+  }
+
+  async function handleEdit(id) {
+    if (!session) {
+      alert("You need to be signed in to edit a comment.");
+      return;
+    }
+
+    setEditMode(true);
+
+    try {
+      await axios.put(`/api/comment/${id}`);
+      onNewComment();
+
+      mutate(`/api/comment?postId=${post.id}`, (cachedData) => {
+        return cachedData.filter((comment) => comment.id !== id);
+      });
+    } catch (error) {
+      setEditMode(false);
+      console.log(error);
+      return <div>Error loading comments</div>;
+    }
+
+    if (!comments) {
+      return <div>Loading comments...</div>;
+    }
+
+    return <button type="submit">Submit</button>;
   }
 
   return (
@@ -46,15 +96,31 @@ export default function CommentsSection({ post, onNewComment }) {
         <h1 className="text-xl font-bold mb-2">Comments</h1>
         {post.comments?.length
           ? post.comments.map(({ id, content, user }) => (
-              <Comment
-                key={id}
-                content={content}
-                user={user}
-                onDelete={() => handleDelete(id)}
-              />
+              <div key={id}>
+                <Comment
+                  content={content}
+                  user={user}
+                  deleteBtn={
+                    session &&
+                    user.id === session.user.id && (
+                      <>
+                        <Popup
+                          onDeleteClick={() => handleDelete(id)}
+                          onEditClick={() => setEditMode(true)}
+                          popupIcon={
+                            <CiMenuKebab className="cursor-pointer text-gray-500 hover:text-gray-700 text-2xl" />
+                          }
+                        />
+                      </>
+                    )
+                  }
+                  showInput={editMode}
+                  onSubmit={() => handleEdit(id)}
+                />
+              </div>
             ))
           : "No comments yet :("}
-        {showForm ? (
+        {session ? (
           <>
             <form
               className="flex flex-col border-t-2 mt-4"
